Remove dashbpard version from models
